Alber Charity Project (PHP + MYSQL)
Al Ber Charity is a project for students at Taibah University to serve charities and help them reach cases in need of care and to the poor. Through the site, the Association provides the clothes and food required for those cases. You as an ordinary citizen can donate through the site the money, clothes and food that you can